Rename to ObjectTypes.h
