+++ src/openrct2/object/ObjectTypes.h
